requirements.txt
absl-py==1.4.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==3.7.1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.7.0
blis==1.0.1
branca==0.7.2
breadability==0.1.20
build==1.2.1
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.5.7
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.1.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.6
cloudpathlib==0.20.0
coloredlogs==15.0.1
comm==0.2.1
confection==0.1.5
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.8
dataclasses-json==0.5.14
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
dnspython==2.5.0
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.104.0
fastjsonschema==2.19.1
filelock==3.12.2
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==23.5.26
folium==0.17.0
fonttools==4.41.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
future==1.0.0
gast==0.4.0
gensim==4.3.1
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.19.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
gps==3.19
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.4
imbalanced-learn==0.12.2
img2table==1.2.10
importlib-metadata==7.0.0
importlib_resources==6.4.0
InstructorEmbedding==1.0.1
ipykernel==6.29.2
ipython==8.21.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jiter==0.8.2
joblib==1.3.1
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==2.13.1rc0
kiwisolver==1.4.4
kubernetes==29.0.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langcodes==3.4.1
langsmith==0.2.4
language_data==1.2.0
libclang==16.0.0
llama-index==0.10.26
llama-index-agent-openai==0.2.1
llama-index-cli==0.1.11
llama-index-core==0.10.26
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-instructor==0.1.3
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-ollama==0.1.2
llama-index-llms-openai==0.1.14
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.13
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-chroma==0.1.6
llama-parse==0.4.0
llamaindex-py-client==0.1.15
llvmlite==0.42.0
lxml==4.9.3
marisa-trie==1.2.1
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
minijinja==1.0.16
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.12.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.59.0
numexpr==2.8.4
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.58.1
openapi-schema-pydantic==1.2.4
opencv-contrib-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
opt-einsum==3.3.0
orjson==3.10.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandoc==2.4
pandocfilters==1.5.1
parso==0.8.3
pdf2image==1.17.0
pexpect==4.9.0
Pillow==10.0.0
platformdirs==4.2.0
plotly==5.24.1
plumbum==1.9.0
ply==3.11
polars==0.20.7
posthog==3.5.0
preshed==3.0.9
prometheus_client==0.21.0
prompt-toolkit==3.0.43
protobuf==4.23.2
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.17.2
pymongo==4.6.1
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pyodbc==5.2.0
pyparsing==3.0.9
pypdf==4.0.1
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.2
ratelim==0.1.6
referencing==0.35.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rocket-python==1.3.4
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.22.0
Send2Trash==1.8.3
sentence-transformers==2.6.1
sentencepiece==0.1.99
shellingham==1.5.4
shutup==0.2.0
six==1.16.0
sklearn==0.0.post7
smart-open==6.3.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.19
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
streamlit==1.39.0
striprtf==0.0.26
sumy==0.11.0
sympy==1.12
tabula==1.0.5
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorflow==2.13.0rc0
tensorflow-estimator==2.13.0rc0
tensorflow-macos==2.13.0rc0
termcolor==2.3.0
terminado==0.18.1
tesseract==0.1.3
thinc==8.3.2
threadpoolctl==3.2.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.15.2
toml==0.10.2
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.39.2
trio==0.25.1
trio-websocket==0.11.1
tweet-preprocessor==0.6.0
typer==0.12.0
typer-cli==0.12.0
typer-slim==0.12.0
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.19.0
virtualenv==20.25.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.2
widgetsnbextension==4.0.13
wrapt==1.14.1
wsproto==1.2.0
xgboost==2.1.2
XlsxWriter==3.1.9
xyzservices==2024.6.0
yarl==1.9.2
zipp==3.18.1
